Real first draft of build llvm in batch
Created two docker container images, one for building agains debian and
one for building against alpine linux. Wrote a batch pipeline that
builds them both, and installs all the tools and llvm-lit (which is not
normally installed, but _should be_).

In the interest of having a consistent compiler in any eventual
image/CI, I also build clang and lld.

Furthermore, the docker images and batch script use [sccache]
a distributed compilation cache (like ccache but it supports cloud
object storage). So it's possible that an eventual ci pipeline would be
able to 'build' LLVM on every PR.

[sccache]: https://github.com/mozilla/sccache